refactor(gpth/dom): Rename `handleElements` to `dom` and fix notification issues
- Rename `handleElements` file to `dom` for better clarity and update all references across the codebase
- Fix the broken notification closing functionality to ensure proper behavior
- Increase the display duration for notifications when displaying longer error messages for better readability
- Ensure seamless integration and improved user experience with the updated notification system

Changes summary:
Renamed `handleElements` to `dom`, fixed the broken notification closing functionality, and increased the display duration for longer error messages.
