+++ src/sass/gpthemes/theme-manager/components/_notification.scss
@@ -32,7 +32,7 @@
             padding-right: calc(var(--p) * 1.5);
         }
 
-        &__close {
+        &__close-btn {
             position: absolute;
             top: 2px;
             right: 4px;
